requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.46.0
bottle==0.13.3
bottle-websocket==0.2.9
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
Eel==0.18.1
future==1.0.0
gevent==25.4.2
gevent-websocket==0.10.1
greenlet==3.2.2
idna==3.10
iso8601==2.1.0
packaging==25.0
pefile==2023.2.7
pycparser==2.22
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pyparsing==3.2.3
pyserial==3.5
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.3
setuptools==80.4.0
typing_extensions==4.13.2
urllib3==2.4.0
zope.event==5.0
zope.interface==7.2
